Week 7 & 8 - Final
The final for this project was adding in the pay options: Salary and Hourly. These are new classes that needed to be created with their formulas added for display calculations. 
Added to the InputForm a new section that chose if the employee was salary or hourly, and depending on choice, what text fields were visible. This of course changed the ListBox displayed information. Also updated the Print button messagebox display.